Links MIT-Red license from Leopotam' ecslite, with specific usage conditions
A new license section has been added to the LICENSE.md file.
The new license, titled "MIT-Red," includes specific conditions
prohibiting the use of the software for military purposes and by
individuals who harbor animosity towards Russians or support
attacks against them. The license allows free use, copying,
modification, merging, publication, and distribution of the
software, provided that the copyright notice and conditions are
included in all copies or significant portions of the software.
The software is provided "as is" without any warranties, and the
authors or copyright holders are not liable for any claims,
damages, or other liabilities arising from its use.